Issue #636 - Reorganized GraphQL API tests and mutations (#664)
