+++ packages/gtmapi/lmsrvlabbook/api/mutations/datasetfiles.py
@@ -275,4 +275,3 @@ class MakeDatasetDirectory(graphene.ClientIDMutation):
 
         return MakeDatasetDirectory(new_dataset_file_edge=DatasetFileConnection.Edge(node=DatasetFile(**create_data),
                                                                                      cursor=cursor))
-
