+++ packages/gtmapi/lmsrvlabbook/api/mutations/migrations.py
@@ -24,8 +24,6 @@ import graphene
 from gtmcore.logging import LMLogger
 
 from gtmcore.inventory.inventory import InventoryManager
-from gtmcore.inventory.branching import BranchManager
-from gtmcore.activity import ActivityStore, ActivityDetailRecord, ActivityDetailType, ActivityRecord, ActivityType
 from gtmcore.workflows import LabbookWorkflow, MergeOverride
 
 from lmsrvcore.auth.user import get_logged_in_username, get_logged_in_author
