+++ packages/gtmapi/lmsrvlabbook/tests/test_labbook_remote_ops.py
@@ -93,76 +93,6 @@ DUMMY_DATA = [
 
 class TestLabBookRemoteOperations(object):
 
-    def test_delete_remote_labbook_dryrun(self, fixture_working_dir):
-        """Test deleting a LabBook on a remote server - dry run"""
-
-        delete_query = f"""
-        mutation delete {{
-            deleteRemoteLabbook(input: {{
-                owner: "default",
-                labbookName: "new-labbook",
-                confirm: false
-            }}) {{
-                success
-            }}
-        }}
-        """
-        r = fixture_working_dir[2].execute(delete_query)
-        print(r)
-        assert 'errors' not in r
-        assert r['data']['deleteRemoteLabbook']['success'] is False
-
-    @responses.activate
-    def test_delete_remote_labbook(self, fixture_working_dir):
-        """Test deleting a LabBook on a remote server"""
-        # Setup responses mock for this test
-        responses.add(responses.GET, 'https://usersrv.gigantum.io/key',
-                      json={'key': 'afaketoken'}, status=200)
-        responses.add(responses.GET, 'https://repo.gigantum.io/api/v4/projects/default%2Fnew-labbook',
-                      json=[{
-                              "id": 27,
-                              "description": "",
-                            }],
-                      status=200)
-        responses.add(responses.DELETE, 'https://repo.gigantum.io/api/v4/projects/default%2Fnew-labbook',
-                      json={
-                                "message": "202 Accepted"
-                            },
-                      status=202)
-        responses.add(responses.GET, 'https://repo.gigantum.io/api/v4/projects/default%2Fnew-labbook',
-                      json=[{
-                                "message": "404 Project Not Found"
-                            }],
-                      status=404)
-        responses.add(responses.DELETE, 'https://api.gigantum.com/read/index/default%2Fnew-labbook',
-                      json=[{
-                                "message": "success"
-                            }],
-                      status=204)
-
-
-        im = InventoryManager(fixture_working_dir[0])
-        lb = im.create_labbook("default", "default", "new-labbook")
-
-        delete_query = f"""
-        mutation delete {{
-            deleteRemoteLabbook(input: {{
-                owner: "default",
-                labbookName: "new-labbook",
-                confirm: true
-            }}) {{
-                success
-            }}
-        }}
-        """
-
-        r = fixture_working_dir[2].execute(delete_query)
-        assert 'errors' not in r
-        assert r['data']['deleteRemoteLabbook']['success'] is True
-
-        # Try deleting again, which should return an error
-        r2 = fixture_working_dir[2].execute(delete_query)
-        assert 'errors' in r2
 
     def test_list_remote_labbooks_invalid_args(self, fixture_working_dir, snapshot):
         """test list labbooks"""
