+++ packages/gtmcore/gtmcore/dispatcher/tests/test_dispatcher.py
@@ -169,129 +169,6 @@ class TestDispatcher(object):
         assert d.query_task(job_ref_1).status == 'failed'
         assert d.query_task(job_ref_2).status == 'deferred'
 
-    def test_build_docker_image(self, temporary_worker, mock_config_file):
-        w, d = temporary_worker
-        erm = RepositoryManager(mock_config_file[0])
-        erm.update_repositories()
-        erm.index_repositories()
-        im = InventoryManager(mock_config_file[0])
-        lb = im.create_labbook('unittester', 'unittester', 'unittest-dispatcher-build-image',
-                               description="Testing docker building.")
-        cm = ComponentManager(lb)
-        cm.add_base(gtmcore.fixtures.ENV_UNIT_TEST_REPO, 'ut-busybox', 0)
-        ib = ImageBuilder(lb)
-        ib.assemble_dockerfile(write=True)
-        assert os.path.exists(os.path.join(lb.root_dir, '.gigantum', 'env', 'Dockerfile'))
-        docker_kwargs = {
-            'path': lb.root_dir,
-            'nocache': True
-        }
-
-        job_ref = d.dispatch_task(bg_jobs.build_labbook_image, kwargs=docker_kwargs)
-
-        elapsed_time = 0
-        while True:
-            status = d.query_task(job_ref).status
-            print(status)
-            r = d.query_task(job_ref)
-            print(r.meta)
-            if elapsed_time > 2:
-                assert r.meta.get('feedback')
-            if status in ['success', 'failed', 'finished']:
-                print(r.exc_info)
-                break
-            if elapsed_time > 60:
-                w.terminate()
-                assert False, "timed out {}".format(status)
-            elapsed_time = elapsed_time + 1
-            time.sleep(1)
-        w.terminate()
-
-        res = d.query_task(job_ref)
-        assert res
-        assert res.status == 'finished'
-
-    def test_start_and_stop_docker_container(self, temporary_worker, mock_config_file):
-        # start_docker_container(docker_image_id, exposed_ports, volumes_dict) -> str:
-        w, d = temporary_worker
-
-        erm = RepositoryManager(mock_config_file[0])
-        erm.update_repositories()
-        erm.index_repositories()
-
-        # Create a labbook
-        lb = InventoryManager(mock_config_file[0]).create_labbook('unittester', 'unittester',
-                                                                  'unittest-start-stop-job',
-                                                                  description="Testing docker building.")
-        cm = ComponentManager(lb)
-        cm.add_base(gtmcore.fixtures.ENV_UNIT_TEST_REPO, 'quickstart-jupyterlab', 2)
-
-        ib = ImageBuilder(lb)
-        ib.assemble_dockerfile(write=True)
-
-        docker_kwargs = {
-            'path': lb.root_dir,
-            'nocache': True,
-            'username': 'unittester'
-        }
-
-        client = get_docker_client()
-        img_list = client.images.list()
-
-        try:
-            from gtmcore.container.utils import infer_docker_image_name
-            owner = InventoryManager().query_owner(lb)
-            client.images.remove(infer_docker_image_name(labbook_name=lb.name, owner=owner,
-                                                         username='unittester'))
-        except:
-            pass
-
-        m = {'method': 'build_image',
-             'labbook': "unittest-start-stop-job"}
-
-        job_ref = d.dispatch_task(bg_jobs.build_labbook_image,
-                                  kwargs=docker_kwargs, metadata=m)
-
-        j = d.query_task(job_ref)
-        assert hasattr(j, 'meta')
-        assert j.meta.get('labbook') == "unittest-start-stop-job"
-
-        elapsed_time = 0
-        while True:
-            status = d.query_task(job_ref).status
-            print(status)
-            if status in ['success', 'failed', 'finished']:
-                print(d.query_task(job_ref).exc_info)
-                break
-            if elapsed_time > 60:
-                w.terminate()
-                assert False, "timed out {}".format(status)
-            elapsed_time = elapsed_time + 1
-            time.sleep(1)
-
-        res = d.query_task(job_ref)
-        assert res
-        print(res.status)
-        assert res.status == 'finished'
-
-        # Finish building image
-
-        startc_kwargs = {
-            'root': lb.root_dir,
-            'config_path': lb.client_config.config_file,
-            'username': 'unittester'
-        }
-        # Start the docker container, and then wait till it's done.
-        container_id = bg_jobs.start_labbook_container(**startc_kwargs)
-        time.sleep(5)
-        assert get_docker_client().containers.get(container_id).status == 'running'
-
-        # Stop the docker container, and wait until that is done.
-        print(container_id)
-        bg_jobs.stop_labbook_container(container_id)
-
-        w.terminate()
-
     def test_simple_scheduler(self, temporary_worker, mock_config_file):
         # Run a simple tasks that increments the integer contained in a file.
         w, d = temporary_worker
