+++ packages/gtmcore/gtmcore/inventory/tests/test_inventory_labbooks.py
@@ -1,7 +1,4 @@
-import shutil
-
 import pytest
-import getpass
 import os
 import shutil
 import yaml
